change after evaluation

+++ getData.php
@@ -1,4 +1,11 @@
 <?php
+session_start();
+if(!isset($_SESSION['loggedin']) || $_SESSION['loggedin'] !==true)
+{
+    header("location: login.php");
+}
+?>
+<?php
 
 include 'connect.php';
 
@@ -16,17 +23,32 @@ while( $row=mysqli_fetch_array($result) ){
     $state=$row['state'];
     $ctc_ug=$row['ctc_ug'];
 
-    $html .='<tr>';
-    $html .='<td>'.$cname.'</td>';
-    $html .='<td>'.$cname_abb.'</td>';
-    $html .='<td>'.$city.'</td>';
-    $html .='<td>'.$state.'</td>';
-    $html .='<td>'.$ctc_ug.'</td>';
-    $html .='<td>
-    <button class="btnd"><a href="delete.php?deletecname='.$cname.' & deletecity='.$city.'" ><i class="fa fa-trash" style="color:red"></i></a></button>
-    <button class="btnu"><a href="update.php?updatecname='.$cname.' & updatecity='.$city.'" ><i class="fa fa-edit" style="color:green"></i></a></button>
-    </td>';
-    $html .='</tr>';
+    if($_SESSION["user_role"]=="TPO")
+    {
+        $html .='<tr onclick="window.location.href=\'view.php?updatecname='.$cname.' & updatecity='.$city.'\'" style = "cursor:pointer">';
+        $html .='<td>'.$cname.'</td>';
+        $html .='<td>'.$cname_abb.'</td>';
+        $html .='<td>'.$city.'</td>';
+        $html .='<td>'.$state.'</td>';
+        $html .='<td>'.$ctc_ug.'</td>';
+        $html .='<td>
+        <button class="btnd"><a href="delete.php?deletecname='.$cname.' & deletecity='.$city.'" ><i class="fa fa-trash" style="color:red"></i></a></button>
+        <button class="btnu"><a href="update.php?updatecname='.$cname.' & updatecity='.$city.'" ><i class="fa fa-edit" style="color:green"></i></a></button>
+        </td>';
+        $html .='</tr>';
+    }
+    else
+    {
+        $html .='<tr onclick="window.location.href=\'view.php?updatecname='.$cname.' & updatecity='.$city.'\'" style = "cursor:pointer">';
+        $html .='<td>'.$cname.'</td>';
+        $html .='<td>'.$cname_abb.'</td>';
+        $html .='<td>'.$city.'</td>';
+        $html .='<td>'.$state.'</td>';
+        $html .='<td>'.$ctc_ug.'</td>';
+        $html .='<td>
+        </td>';
+        $html .='</tr>';
+    }
 }
 
 echo $html;
